refactor: correct component usage
