+++ packages/console/src/components/CreateTenantModal/SelectTenantPlanModal/SkuCardItem/FeaturedSkuContent/index.tsx
@@ -1,9 +1,9 @@
 import classNames from 'classnames';
 
-import Failed from '@/assets/icons/failed.svg';
-import Success from '@/assets/icons/success.svg';
+import Failed from '@/assets/icons/failed.svg?react';
+import Success from '@/assets/icons/success.svg?react';
 
-import * as styles from '../../PlanCardItem/FeaturedPlanContent/index.module.scss';
+import styles from '../../PlanCardItem/FeaturedPlanContent/index.module.scss';
 
 import useFeaturedSkuContent from './use-featured-sku-content';
